ran through and completed friend system on backend

+++ app/controllers/friendships_controller.rb
@@ -3,46 +3,35 @@ class FriendshipsController < ApplicationController
 
 	# For when a user clicks ACCEPT on a friend request, create new friendship
 	def create
-		friendship_one = Friendship.create!(friendship_params)
+		user = User.find(friendship_params[:user_id])
+		friend = User.find(friendship_params[:friend_id])
 
-		# Create a second friendship with reverse attributes and link to first friendship
-		friendship_two = Friendship.create!(
-			user_id: params[:friend_id],
-			friend_id: params[:user_id],
-			corresponding_friendship_id: friendship_one.id
-		)
+		# can't be friends with yourself
+		if user.id == friend.id
+			render json: { error: "You can't be friends with yourself" }
+		else
+			friendship_one = Friendship.create!(friendship_params)
+			# Create a second friendship with reverse attributes and link to first friendship
+			friendship_two = Friendship.create!(
+				user: friend,
+				friend: user,
+				corresponding_friendship: friendship_one
+			)
+			# Link first friendship to second
+			friendship_one.update!(corresponding_friendship: friendship_two)
 
-		# Link first friendship to second
-		friendship_one.update!(corresponding_friendship_id: friendship_two.id)
+			# Destroy corresponding friend request
+			friend_request = FriendRequest.find_by(receiver: user, requester: friend)
+			friend_request.destroy
 
-		# Destroy corresponding friend request
-		friend_request = FriendRequest.find_by(
-			receiver_id: params[:user_id],
-			requester_id: params[:friend_id]
-		)
-		friend_request.destroy
-
-		# Render JSON
-		render json: {created: FriendshipSerializer.new(friendship_one), destroyed: FriendRequestSerializer.new(friend_request) }, status: :created
+			# Render JSON
+			render json: { created: FriendshipSerializer.new(friendship_one), destroyed: FriendRequestSerializer.new(friend_request) }, status: :created
+		end
 	end
 
 	def my_friends
-		user = User.find(params[:user_id])
+		user = User.find(friendship_params[:user_id])
 		render json: user.friends
-	end
-
-	def friend_status
-		# Send back message depending on whether 1. they haven't responded to your friend request or 2. you haven't responded to them
-		# Check whether or not friends
-		if FriendRequest.find_by(requester: params[:user_id], receiver: params[:friend_id])
-			render json: { message: "Pending Them" }
-		elsif FriendRequest.find_by(requester: params[:friend_id], receiver: params[:user_id])
-			render json: { message: "Pending You" }
-		elsif @friendship
-			render json: { message: "Friends" }
-		else
-			render json: { message: "Not Friends" }
-		end
 	end
 
 	def destroy
@@ -59,6 +48,6 @@ class FriendshipsController < ApplicationController
 	end
 
 	def set_friendship
-		@friendship = Friendship.find_by(user_id: params[:user_id], friend_id: params[:friend_id])
+		@friendship = Friendship.find(params[:id])
 	end
 end
